requirements.txt
dj-database-url==0.5.0
Django==1.11
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
Pillow==6.1.0
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.2
six==1.12.0
whitenoise==4.1.4
